Check before attempting to remove composer from phar archive

+++ build/compile.php
@@ -10,5 +10,9 @@ if(file_exists('emulator.phar')) {
 $phar = new Phar('emulator.phar', 0, 'emulator.phar');
 
 $phar->buildFromDirectory(dirname(__DIR__));
-$phar->delete('composer.phar');
+
+if(file_exists('phar://emulator.phar/composer.phar')) {
+  $phar->delete('composer.phar');
+}
+
 $phar->setDefaultStub('emulator/emulator.php');
